AiInternTask/Task1/requirements.txt
# requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv

# Database and vector search
SQLAlchemy
faiss-cpu
numpy
openai
sentence-transformers

# For potential future use
# requests